Added Narrator and Dialogue
Added a Narrator scriptable script that takes N string of lines where N is the number of lines inserted.
Added a NarratorTrigger class with a function named TriggerNarrator that will call StartNarrator take a Narrator object, and start the dialogue depending on the button pressed.
Added a NarratorManager MonoBehaviour class that starts the dialogue and displays it in a text box. It waits for a certain time until it updates and displays another line.